+++ umbra/components/factory/scriptEditor/models.py
@@ -254,7 +254,7 @@ class ProjectsModel(umbra.ui.models.GraphModel):
 		:return: EditorNode nodes. ( List )
 		"""
 
-		return self.findFamily("EditorNode", node=node or self.__defaultProjectNode)
+		return self.findFamily("Editor", node=node or self.__defaultProjectNode)
 
 	def listFileNodes(self, node=None):
 		"""
@@ -264,7 +264,7 @@ class ProjectsModel(umbra.ui.models.GraphModel):
 		:return: FileNode nodes. ( List )
 		"""
 
-		return self.findFamily("FileNode", node=node or self.__defaultProjectNode)
+		return self.findFamily("File", node=node or self.__defaultProjectNode)
 
 	def listDirectoryNodes(self):
 		"""
@@ -273,7 +273,7 @@ class ProjectsModel(umbra.ui.models.GraphModel):
 		:return: DirectoryNode nodes. ( List )
 		"""
 
-		return self.findFamily("DirectoryNode")
+		return self.findFamily("Directory")
 
 	def listProjectNodes(self, ignoreDefaultProjectNode=True):
 		"""
@@ -283,7 +283,7 @@ class ProjectsModel(umbra.ui.models.GraphModel):
 		:return: ProjectNode nodes. ( List )
 		"""
 
-		projectNodes = self.findFamily("ProjectNode")
+		projectNodes = self.findFamily("Project")
 		return filter(lambda x: x != self.__defaultProjectNode, projectNodes) \
 		if ignoreDefaultProjectNode else projectNodes
 
@@ -703,7 +703,7 @@ class ProjectsModel(umbra.ui.models.GraphModel):
 			else:
 				parentNode = foundations.common.getFirstItem(
 							[node for node in foundations.walkers.nodesWalker(rootNode) \
-							if node.family == "DirectoryNode" and node.path == parentDirectory])
+							if node.family == "Directory" and node.path == parentDirectory])
 
 			if not parentNode:
 				continue
@@ -750,9 +750,9 @@ class ProjectsModel(umbra.ui.models.GraphModel):
 		"""
 
 		for node in reversed(list(foundations.walkers.nodesWalker(node))):
-			if node.family == "DirectoryNode":
+			if node.family == "Directory":
 				self.unregisterDirectory(node)
-			elif node.family == "FileNode":
+			elif node.family == "File":
 				self.unregisterFile(node)
 
 	def updateProjectNodes(self, node):